# api/openapi/v1/auth.yaml
---
openapi: 3.0.0

info:
  title: 'auth'
  version: 1.0.0

servers:
  - url: 'http://localhost:8080/api/v1'
    description: local

security: []

paths:
  /auth/signup:
    post:
      tags:
        - auth
      summary: Create a new user
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/AuthenticationUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedResponse'
  /auth/login:
    post:
      tags:
        - auth
      summary: Login an existing user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successfully logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tokens'
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: User inactive
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/AuthenticationUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedResponse'
  /auth/refresh:
    post:
      tags:
        - auth
      summary: Rotate the session refresh token
      description: Refresh tokens are single-use. Reusing one ends that session, including successor tokens.
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tokens'
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid, expired, reused or inactive session
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/AuthenticationUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedResponse'
  /auth/logout:
    post:
      tags:
        - auth
      summary: End the session
      description: The refresh token identifies the session. Repeated logout succeeds, including after access-token expiry.
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Tokens invalidated successfully
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid refresh token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/AuthenticationUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedResponse'
  /users/me:
    get:
      tags:
        - users
      summary: Retrieve current user information
      operationId: users.me.read
      x-required-permissions:
        - users:read_self
      security:
        - jwt: []
        - apiKey: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/UnexpectedResponse'
    put:
      tags:
        - users
      summary: Update current user credentials
      description: Requires the current password. Changing email or password ends all existing JWT sessions.
      operationId: users.me.update
      x-required-permissions:
        - users:update_self
      security:
        - jwt: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSelfRequest'
      responses:
        '200':
          description: User updated
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/AuthenticationUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedResponse'
  /users:
    get:
      tags:
        - users
      summary: List users
      operationId: users.list
      x-required-permissions:
        - users:list
      security:
        - jwt: []
        - apiKey: []
      parameters:
        - name: limit
          in: query
          description: Number of users to return. Omit or use 0 for the default of 10. Empty values are invalid.
          schema:
            type: integer
            default: 10
            minimum: 0
            maximum: 100
        - name: offset
          in: query
          description: Number of users to skip. Omit for the default of 0. Empty values are invalid.
          schema:
            type: integer
            default: 0
            minimum: 0
            maximum: 2147483647
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: Invalid pagination
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/UnexpectedResponse'
    post:
      tags:
        - users
      summary: Create a new user
      operationId: users.create
      x-required-permissions:
        - users:create
      security:
        - jwt: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/UnexpectedResponse'
  /users/{uuid}:
    get:
      tags:
        - users
      summary: Get user by UUID
      operationId: users.get
      x-required-permissions:
        - users:read_others
      security:
        - jwt: []
        - apiKey: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid UUID
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/UnexpectedResponse'
    put:
      tags:
        - users
      summary: Update user
      operationId: users.update
      x-required-permissions:
        - users:update
      security:
        - jwt: []
        - apiKey: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/UnexpectedResponse'
    delete:
      tags:
        - users
      summary: Delete user
      operationId: users.delete
      x-required-permissions:
        - users:delete
      security:
        - jwt: []
        - apiKey: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted
        '400':
          description: Invalid UUID
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/UnexpectedResponse'
# ---
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
  responses:
    RateLimited:
      description: Too many authentication attempts
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    AuthenticationUnavailable:
      description: Authentication storage or rate limiter is unavailable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    UnexpectedResponse:
      description: Unexpected response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          description: A URI reference that identifies the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          description: The HTTP status code generated by the origin server.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem.
        errors:
          type: array
          items:
            type: object
          description: Optional list of additional error details.
    User:
      type: object
      properties:
        uuid:
          type: string
          default: ""
        email:
          type: string
          default: ""
        created_at:
          type: string
          default: ""
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
    Tokens:
      type: object
      properties:
        access_token:
          type: string
          default: ""
        refresh_token:
          type: string
          default: ""
        expires_in:
          type: integer
          default: 0
    SignUpRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: Trimmed, lowercased, then validated as an email address by the service.
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: >-
            At least 8 Unicode characters and at most 72 UTF-8 bytes. Used exactly as supplied, including whitespace. New
            passwords must also meet the configured strength requirement.
          example: "12#$abCD%$"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: Trimmed, lowercased, then validated as an email address by the service.
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: >-
            At least 8 Unicode characters and at most 72 UTF-8 bytes. Used exactly as supplied, including whitespace.
          example: "12#$abCD%$"
    RefreshRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          default: ""
    LogoutRequest:
      type: object
      required:
        - refresh_token
      properties:
        access_token:
          description: Optional; logout uses the refresh token to identify the session.
          type: string
          default: ""
        refresh_token:
          type: string
          default: ""
    CreateUserRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: Trimmed, lowercased, then validated as an email address by the service.
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: >-
            At least 8 Unicode characters and at most 72 UTF-8 bytes. Used exactly as supplied, including whitespace. New
            passwords must also meet the configured strength requirement.
          example: "12#$abCD%$"
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          description: Trimmed, lowercased, then validated as an email address by the service.
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: >-
            At least 8 Unicode characters and at most 72 UTF-8 bytes. Used exactly as supplied, including whitespace. New
            passwords must also meet the configured strength requirement.
          example: "12#$abCD%$"
    UpdateSelfRequest:
      type: object
      description: current_password is required when email or password is supplied.
      properties:
        current_password:
          type: string
          minLength: 8
          maxLength: 72
          description: Existing password, used exactly as supplied; at most 72 UTF-8 bytes.
          writeOnly: true
        email:
          type: string
          description: Trimmed, lowercased, then validated as an email address by the service.
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: >-
            At least 8 Unicode characters and at most 72 UTF-8 bytes. Used exactly as supplied, including whitespace. New
            passwords must also meet the configured strength requirement.
          example: "12#$abCD%$"